Creating v2.0 with QC workflow, annotations, and crash handling
Intended to run in Blender 2.93, but works in any 2.90x generally. QC workflow allows auto and manually 'tagging' of issues related to specific files. Implements reading and outputs to text file to detect if blender has crashed, with the intent that this file is ran by a parent auto-restart wrapper.